requirements.txt
Faker~=19.6.1
pytest~=7.4.2
pytest-ordering~=0.6
selenium~=4.12.0
ruff~=0.0.289
black~=23.9.1
requests~=2.31.0
allure-pytest~=2.13.2
